moving some http streaming files around, reorganizing package
