+++ colossus/src/main/scala/colossus/protocols/http/streaming/package.scala
@@ -1,0 +1,9 @@
+package colossus.protocols.http
+
+import colossus.streaming.Source
+
+package object streaming {
+
+  implicit object StreamingHttpRequestOps extends MessageOps[HttpRequestHead, Source[Data], StreamingHttpRequest](StreamingHttpRequest.apply _)
+
+}
